consumer-python/requirements.txt
kafka-python==2.1.2
kafka-python
pandas
openpyxl


